ACU:660: Add mosaico plugin support to versafix-compucorp template
For the template to support new plugin, we override civicrm/mosaico/iframe route in mosaico extension and add a smarty variable that has the url to our plugin script. we also added a new widget called fontLoader to support loading of fonts from javascript file instead of the template file, the next work would be getting this fontLoader widget to load font from php/DB into the javascript file.